Archive Directory
Added an archive directory that images can be moved to after copying wanted images to another location. This allows you to copy the images you want to keep somewhere else, and then move all of the images out of the destination directory into the archive directory, so that newly downloaded images are easy to find.